+++ src/proxy/routerGen.go
@@ -139,6 +139,10 @@ func RouterGen(route utils.ProxyRouteConfig, router *mux.Router, destination htt
 		destination = utils.BlockPostWithoutReferer(destination)
 	}
 
+	if !route.DisableHeaderHardening {
+		destination = utils.SetSecurityHeaders(destination)
+	}
+
 	destination = tokenMiddleware(route.AuthEnabled, route.AdminOnly)(utils.CORSHeader(originCORS)((destination)))
 
 	origin.Handler(destination)