+++ src/utils/utils.go
@@ -572,3 +572,11 @@ func GetNetworkUsage() NetworkStatus {
 
 	return NetworkStatus{}
 }
+
+func GetClientIP(req *http.Request) string {
+	/*ip := req.Header.Get("X-Forwarded-For")
+	if ip == "" {
+		ip = req.RemoteAddr
+	}*/
+	return req.RemoteAddr
+}